pyproject.toml
[tool.poetry]
name = "doubledistill"
version = "0.1.0"
description = ""
authors = ["Don Dennis <dondennis@cmu.edu>"]
readme = "README.md"
packages =[
    {include = "ddist", from = "doubledistill"},
    {include = "ddistexps", from = "doubledistill"},
]


[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.2"
torchvision = "^0.16.2"
rich = "^13.7.0"
grpcio = "^1.60.0"
pandas = "^1.5.0"
open-clip-torch = "^2.24.0"
deepspeed = "^0.13.2"

[tool.poetry.group.dev.dependencies]
ray = "^2.8.1"
mlflow = "^2.9.2"
python-dotenv = "^1.0.0"

#[tool.poetry.scripts]
#federated_run = "app.federated_model3d_train:main"
#federated_sweep = "app.federated_model3d_train:run_sweep"
#"doubledistill.distexps.baseline.main.py  --expname debug"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"